First pass to support events in liboro
Most of the architecture is in place. What is missing is the
passive watching of the socket for incoming events. Currently events
are only read when a request is send to oro-server, ie, when IConnector::
execute() is called.

Summary of the architecture:
The user register a new event with Ontology::registerEvent()
It must pass an object that implements the OroEventObserver interface that
declares a functor.

When an event (identified by an unique ID) is triggered by oro-server,
SocketConnector::read() call evtCallback() in Ontology that in turn call
the liboro user OroEventObserver functor.